requirements.txt
Django==4.1
djangorestframework==3.14.0
psycopg2==2.9.6
drf-yasg==1.21.5